requirements.txt
blinker==1.9.0
click==8.3.3
flask==3.1.3
flask-cors==5.0.0
gunicorn==25.3.0
itsdangerous==2.2.0
jinja2==3.1.6
libsql==0.1.11
markupsafe==3.0.3
packaging==26.2
python-dotenv==1.2.2
werkzeug==3.1.8
